test: move custom mapper test in a separate file
As suggested, move custom mapper tests in a separate file. Because
AbstractProcessInstanceAssertTest is too big
