+++ qa/abstracts/src/main/java/io/camunda/zeebe/process/test/qa/abstracts/assertions/AbstractProcessInstanceAssertTest.java
@@ -1439,47 +1439,4 @@ public abstract class AbstractProcessInstanceAssertTest {
           .hasVariableWithValue(ProcessPackLoopingServiceTask.TOTAL_LOOPS, "3");
     }
   }
-
-  @Nested
-  class CustomMapperTests {
-
-    private ZeebeClient client;
-    private ZeebeTestEngine engine;
-
-    private final ObjectMapper objectMapper = configureObjectMapper();
-
-    @Test
-    public void shouldDeserializeDateVariables() throws InterruptedException, TimeoutException {
-      final Map<String, Object> variables = new HashMap<>();
-      variables.put("stringProperty", "stringValue");
-      variables.put("numberProperty", 123);
-      variables.put("booleanProperty", true);
-      variables.put("complexProperty", Arrays.asList("Element 1", "Element 2"));
-      variables.put("nullProperty", null);
-      variables.put("javaDate", LocalDateTime.of(2023, 8, 14, 16, 0, 0));
-      variables.put("stringDate", "\"2023-08-14T16:00:00\"");
-      // given
-      Utilities.deployResource(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessInstanceEvent instanceEvent =
-          Utilities.startProcessInstance(
-              engine, client, ProcessPackLoopingServiceTask.PROCESS_ID, variables);
-
-      // then
-      assertThatNoException()
-          .isThrownBy(
-              () ->
-                  variables.forEach(
-                      (key, value) ->
-                          BpmnAssert.assertThat(instanceEvent).hasVariableWithValue(key, value)));
-    }
-
-    private ObjectMapper configureObjectMapper() {
-      final ObjectMapper objectMapper = new ObjectMapper();
-      objectMapper.registerModule(new JavaTimeModule());
-      objectMapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
-      return objectMapper;
-    }
-  }
 }
